Add manual trace. Add Seq to `Aspire.Hosting`.
